remove the last src/run.ts
