+++ .changeset/real-shirts-cry.md
@@ -1,0 +1,5 @@
+---
+'@backstage/cli-node': patch
+---
+
+Internal refactor
