+++ .changeset/slow-horses-destroy.md
@@ -1,0 +1,6 @@
+---
+'@backstage/plugin-catalog-backend-module-incremental-ingestion': patch
+'@backstage/plugin-kubernetes-backend': patch
+---
+
+Create a `dev/index.ts` entrypoint for `yarn start`
